fix: Fix browser authentication issues during discovery (#57)
In a previous [PR](https://github.com/MeltanoLabs/tap-snowflake/pull/53)
of mine I tried to fix this
[issue](https://github.com/MeltanoLabs/tap-snowflake/issues/48), but
while using the tap again I noticed that the issue had not been fixed.
The browser authentication worked in my local execution because I was
using token caching and had the oauth token cached in my local machine
when I ran the tests.

In order to properly fix it I tried multiple approaches, but it was
difficult to pinpoint the exact place where the `stdout` redirection had
to occur for this to work, so I decided to go with a different approach
and replace the `connect` method by a wrapped `connect` method where the
`stdout` is redirected to `stderr` if the authentication type is browser
authentication.

Works like a charm for me, so decided to contribute!